+++ src/main/java/load/datapool/todo/rest/TodoRestController.java
@@ -60,6 +60,9 @@ public final class TodoRestController {
         if (!lockerService.poolExist(env, pool))
             return tableNotFindResponse(fullPoolName);
 
+        if (lockerService.isMarkedAsEmpty(env, pool))
+            return tableIsEmptyResponse(fullPoolName);
+
         try {
             sq = jdbcOperations.queryForObject("select nextval (?)", new String[]{getSeqPrefix(env, pool) + "_rid"}, Long.class);
             ResponseEntity<String> res = ResponseEntity.ok(jdbcOperations.queryForObject("select rid, text, locked from " + fullPoolName + " where rid = ? and locked = false limit 1",
@@ -72,35 +75,15 @@ public final class TodoRestController {
             }
             exp.increaseLatency(env, pool, "get-next-value", start);
             return res;
-
         } catch (EmptyResultDataAccessException e) {
             if (restartPoolLock.tryLock()) {
                 if (fixSequenceState(env, pool, sq)) {
-                    System.out.println("Sequence reseted " + fullPoolName + " last value = " + sq
-//                        + ". Retry count = " + retryGetNextCount
-                    );
-
-                /*
-                todo так как контроллер - это бин, а бин - синглтон, то retryGetNextCount считает сумму рестартов по всем пулам без разбора.
-                Также тут не учитываются временные промежутки.
-                Таким образом, мы просто шлём badRequest без явных оснований.
-                Пока вырезаю эту фичу.
-                 */
-
-//                if (retryGetNextCount >= retryGetNextMaxCount) {
-//                    retryGetNextCount = 0;
-//                    exp.increaseLatency(env, pool, "get-next-value", start);
-//                    return ResponseEntity.badRequest().body(new String(e.getMessage() + "\\\n"
-//                            + extErrText
-//                            + "\\\n Datapool may be empty."
-//                            + "\\\n Very often reset sequense: " + retryGetNextCount
-//                            + " for " + env + "." + pool + " last value = " + sq));
-//                }
-//                retryGetNextCount++;
+                    System.out.println("Sequence reseted for " + fullPoolName + " was value = " + sq);
                 } else {
                     restartPoolLock.unlock();
                     exp.increaseLatency(env, pool, "get-next-value", start);
-                    return ResponseEntity.badRequest().body(new String(e.getMessage() + "\\\n" + extErrText + "\\\n ERROR " + env + "." + pool + " sequence = " + sq));
+//                    return ResponseEntity.badRequest().body(new String(e.getMessage() + "\\\n" + extErrText + "\\\n ERROR " + env + "." + pool + " sequence = " + sq));
+                    return tableIsEmptyResponse(fullPoolName);
                 }
                 restartPoolLock.unlock();
             } else {
@@ -121,33 +104,24 @@ public final class TodoRestController {
     }
 
     private boolean fixSequenceState(String env, String pool, Long currentValue) {
-        String fullPullName = fullPoolName(env, pool);
         if (currentValue < 0) return false; //Some undefined sequence error
         try {
-//            long operation
-//            Long sqMax = jdbcOperations.queryForObject("SELECT nvl(max (rid), 0) FROM " + env + "." + pool + "  where rid >= ? and locked = false", new Object[]{currentValue}, Long.class); //Is exist values forward
             int sqMax = lockerService.firstBiggerUnlockedId(env, pool, currentValue.intValue());
             Long newSqValue = 0L;
             if (sqMax == 0) //Need reset sequence to start, no values forward
             // You can change for sequence CYCLE type if slow puts are planing
             {
-//                long operation
-//                newSqValue = jdbcOperations.queryForObject("SELECT nvl(min (rid), 0) FROM " + env + "." +pool + " where rid > 0 and locked = false",  Long.class);
-                newSqValue = (long) lockerService.firstUnlockRid(env, pool);
+                 newSqValue = (long) lockerService.firstUnlockRid(env, pool);
                 if (newSqValue == 0) {
-                    extErrText += "May be table is empty." + " ";
+                    lockerService.markAsEmpty(env, pool);
                     newSqValue = 1L;
                     jdbcOperations.update("ALTER SEQUENCE " + getSeqPrefix(env, pool) + "_rid" + " RESTART WITH ?", newSqValue);
                     return false;
-                } //May be table is empty
+                }
                 jdbcOperations.update("ALTER SEQUENCE " + getSeqPrefix(env, pool) + "_rid" + " RESTART WITH ?", newSqValue);
                 return true;
-            } else {//need reset sequence next
-                newSqValue = this.jdbcOperations.queryForObject("SELECT nvl(min (rid),0) FROM " + env + "." + pool + " where  locked != true", Long.class);
-                if (newSqValue == 0) {
-                    extErrText += "Some errors when try to make sequence offset." + " ";
-                    return false;
-                } //May be table is empty
+            } else {//need reset sequence to next available value
+                newSqValue = Long.valueOf(sqMax);
                 this.jdbcOperations.update("ALTER SEQUENCE " + getSeqPrefix(env, pool) + "_rid" + " RESTART WITH ?", newSqValue);
                 return true;
             }
@@ -156,7 +130,6 @@ public final class TodoRestController {
             System.err.println(e.getMessage() + "\\\n" + extErrText);
             return false;
         }
-        //return false;
     }
 
     @PostMapping(path = "/put-value")
@@ -199,14 +172,12 @@ public final class TodoRestController {
                                              @RequestParam(value = "rid", defaultValue = "-1") String sRid,
                                              @RequestParam(value = "search-key", defaultValue = "") String searchKey,
                                              @RequestParam(value = "unlock-all", defaultValue = "false") boolean unlockAll) {
-
+        Instant start = Instant.now();
+        exp.increaseRequests(env, pool, "unlock");
         if (!lockerService.poolExist(env, pool))
             return tableNotFindResponse(fullPoolName(env, pool));
 
-        Instant start = Instant.now();
-        exp.increaseRequests(env, pool, "unlock");
         Long rid;
-
         //Check rid valid value
         try {
             rid = Long.parseLong(sRid.trim());  //<-- String to long here
@@ -233,13 +204,12 @@ public final class TodoRestController {
                     }
                 }
             }
-
         } catch (DataAccessException e) {
             exp.increaseLatency(env, pool, "unlock", start);
             return ResponseEntity.badRequest().body(new String(e.getMessage() + "\\\n" + extErrText + "\\\n ERROR " + env + "." + pool));
         }
         exp.increaseLatency(env, pool, "unlock", start);
-        return ResponseEntity.ok(new String("Unlock successfully!"));
+        return ResponseEntity.ok(new String("Unlocked successfully!"));
     }
 
     @GetMapping(path = "/search-by-key")
@@ -355,15 +325,17 @@ public final class TodoRestController {
                     json += "}";
                     this.jdbcOperations.update("insert into " + env + "." + pool + "(rid, text, locked) values (nextval (?),?,?);", getSeqPrefix(env, pool) + "_max", json, false);
                     successCnt++;
-                    //System.out.println(json);
                 }
-
                 // read next line
                 line = reader.readLine();
             }
             reader.close();
 
-            lockerService.putPool(env, pool, successCnt);
+            if (override) {
+                lockerService.putPool(env, pool, successCnt);
+            }else{
+                lockerService.add(env, pool, successCnt);
+            }
 
             if (!notValidRows.equals("")) {
                 exp.increaseLatency(env, pool, "upload-csv-as-json", start);
@@ -379,6 +351,12 @@ public final class TodoRestController {
 
     private ResponseEntity<String> tableNotFindResponse(String poolName) {
         return ResponseEntity.badRequest().body("Table " + poolName + " not found in lockerService");
+    }
+
+    private ResponseEntity<String> tableIsEmptyResponse(String poolName) {
+        return ResponseEntity.badRequest().body("Incorrect result size: expected 1, actual 0\\\n" +
+                "May be table is empty. \\\n" +
+                " ERROR " + poolName);
     }
 
     private String getSeqPrefix(String env, String pool) {
@@ -400,6 +378,8 @@ public final class TodoRestController {
         }
         lockerService.putPool(env, pool);
         try {
+        	
+        	this.jdbcOperations.execute("create schema if not exists "+env+";");
             this.jdbcOperations.execute("" +
                     "create table " + env + "." + pool +
                     "(" +
@@ -444,17 +424,6 @@ public final class TodoRestController {
         }
     }
 
-    private boolean isTableNotFound(Throwable cause) {
-        Pattern pattern = Pattern.compile("Table .* not found");
-        Matcher matcher = pattern.matcher(cause.getMessage());
-
-        if (matcher.find()) {
-            System.out.println("isTableNotFound function = true");
-            return true;
-        }
-        return false;
-    }
-
     private String fullPoolName(String env, String pool) {
         return env + "." + pool;
     }
@@ -475,4 +444,26 @@ public final class TodoRestController {
             return ResponseEntity.badRequest().body("Pool not found");
         }
     }
+    
+    // reset sequence to 1 (zero point of pool)
+    @GetMapping(path = "/resetseq")
+    public ResponseEntity<Object> resetseq(@RequestParam(value = "env", defaultValue = "load") String env,
+                                          @RequestParam(value = "pool", defaultValue = "testpool") String pool) {
+        Instant start = Instant.now();
+        exp.increaseRequests(env, pool, "resetseq");
+        Long newSqValue = 1L;
+
+        try {
+        	jdbcOperations.update("ALTER SEQUENCE " + getSeqPrefix(env, pool) + "_rid" + " RESTART WITH ?", newSqValue);
+        } catch (Exception e) {
+            logger.error(e.getMessage());
+            exp.increaseLatency(env, pool, "resetseq", start);
+            return ResponseEntity.badRequest().body(e.getMessage());
+        }
+        exp.increaseLatency(env, pool, "resetseq", start);
+        return ResponseEntity.ok(new String("Row id sequence has been reseted for pool "+pool));
+    }
+
+    
+    
 }