+++ changelog.md
@@ -1,6 +1,31 @@
 # Changelog
 Service for creating and utilizing test data pools via REST API.
 
+## [1.7.1]
+
+### Changed
+
+- Fixed override=false in /upload-csv-as-json
+- Mark empty pool in locker
+- Fixed next biggest id
+- Check poolExist only in app cache
+
+## [1.6.5]
+
+### Changed
+
+- Fixed next sequence value
+
+## [1.6.4]
+
+### Added
+
+- Resetting datapool sequence has done. /resetseq
+
+### Changed
+
+- Threads properties was optimized
+
 ## [1.6.3]
 
 ### Changed